backend/requirements.txt
fastapi[standard]
pydantic
python-dotenv
sqlmodel
openai==0.28
py-solc-x